Revamp `optional exam 1`
