+++ src/main/kotlin/com/dicoding/exam/optionalexam1/OptionalExam1.kt
@@ -1,0 +1,6 @@
+package com.dicoding.exam.optionalexam1
+
+// TODO
+fun sumOfBigThree(vararg numbers: Int): Int {
+    return 0
+}
