+++ src/main/kotlin/com/dicoding/exam/optionalexam1/OptionalExam1Main.kt
@@ -1,0 +1,8 @@
+package com.dicoding.exam.optionalexam1
+
+private fun main() {
+    println(sumOfBigThree(1, 2, 3, 4, 5, 6, 7, 8, 9, 10) == 27)
+    println(sumOfBigThree(22, 20, 21, 25, 24) == 71)
+    println(sumOfBigThree(100, 1000, 30, 1) == 1130)
+    println(sumOfBigThree(22, 22, 33, 22) == 77)
+}
